Moved PEvent and PNetworkEvent into models folder
