+++ src/shared/models/PNetworkEvent.ts
